ingestion_service/pyproject.toml
[project]
name = "ingestion-service"
version = "0.1.0"
description = "Event ingestion service for the event processing platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "aiokafka",
    "pydantic",
    "pydantic-settings",
    "prometheus-client",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-exporter-otlp",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["ingestion_service"]
